usm: Move USM endpoints to separate files
The change allows us to use linux only definitions for the linux only protocols.
The change introduce an OS specific method to register USM endpoints, which
registers the common endpoints for both linux and windows (http debugging endpoint, telemetry), and
register the linux only protocols (http2, redis, kafka, postgres) and TLS capabilities
(attach, detach, list).

In the following commits and PRs we can re-use structs and definitions that are linux only
to prevent code duplication that can go out of sync.
