+++ cmd/system-probe/modules/network_tracer.go
@@ -20,18 +20,10 @@ import (
 	"sync/atomic"
 	"time"
 
-	coreconfig "github.com/DataDog/datadog-agent/pkg/config/setup"
 	"github.com/DataDog/datadog-agent/pkg/network"
 	networkconfig "github.com/DataDog/datadog-agent/pkg/network/config"
 	"github.com/DataDog/datadog-agent/pkg/network/encoding/marshal"
-	httpdebugging "github.com/DataDog/datadog-agent/pkg/network/protocols/http/debugging"
-	kafkadebugging "github.com/DataDog/datadog-agent/pkg/network/protocols/kafka/debugging"
-	postgresdebugging "github.com/DataDog/datadog-agent/pkg/network/protocols/postgres/debugging"
-	redisdebugging "github.com/DataDog/datadog-agent/pkg/network/protocols/redis/debugging"
-	"github.com/DataDog/datadog-agent/pkg/network/protocols/telemetry"
 	"github.com/DataDog/datadog-agent/pkg/network/tracer"
-	usmconsts "github.com/DataDog/datadog-agent/pkg/network/usm/consts"
-	usm "github.com/DataDog/datadog-agent/pkg/network/usm/utils"
 	"github.com/DataDog/datadog-agent/pkg/system-probe/api/module"
 	sysconfigtypes "github.com/DataDog/datadog-agent/pkg/system-probe/config/types"
 	"github.com/DataDog/datadog-agent/pkg/system-probe/utils"
@@ -151,91 +143,6 @@ func (nt *networkTracer) Register(httpMux *module.Router) error {
 		utils.WriteAsJSON(w, stats, utils.CompactOutput)
 	})
 
-	httpMux.HandleFunc("/debug/http_monitoring", func(w http.ResponseWriter, req *http.Request) {
-		if !coreconfig.SystemProbe().GetBool("service_monitoring_config.enable_http_monitoring") {
-			writeDisabledProtocolMessage("http", w)
-			return
-		}
-		id := utils.GetClientID(req)
-		cs, cleanup, err := nt.tracer.GetActiveConnections(id)
-		if err != nil {
-			log.Errorf("unable to retrieve connections: %s", err)
-			w.WriteHeader(500)
-			return
-		}
-		defer cleanup()
-
-		utils.WriteAsJSON(w, httpdebugging.HTTP(cs.HTTP, cs.DNS), utils.GetPrettyPrintFromQueryParams(req))
-	})
-
-	httpMux.HandleFunc("/debug/kafka_monitoring", func(w http.ResponseWriter, req *http.Request) {
-		if !coreconfig.SystemProbe().GetBool("service_monitoring_config.enable_kafka_monitoring") {
-			writeDisabledProtocolMessage("kafka", w)
-			return
-		}
-		id := utils.GetClientID(req)
-		cs, cleanup, err := nt.tracer.GetActiveConnections(id)
-		if err != nil {
-			log.Errorf("unable to retrieve connections: %s", err)
-			w.WriteHeader(500)
-			return
-		}
-		defer cleanup()
-
-		utils.WriteAsJSON(w, kafkadebugging.Kafka(cs.Kafka), utils.GetPrettyPrintFromQueryParams(req))
-	})
-
-	httpMux.HandleFunc("/debug/postgres_monitoring", func(w http.ResponseWriter, req *http.Request) {
-		if !coreconfig.SystemProbe().GetBool("service_monitoring_config.enable_postgres_monitoring") {
-			writeDisabledProtocolMessage("postgres", w)
-			return
-		}
-		id := utils.GetClientID(req)
-		cs, cleanup, err := nt.tracer.GetActiveConnections(id)
-		if err != nil {
-			log.Errorf("unable to retrieve connections: %s", err)
-			w.WriteHeader(500)
-			return
-		}
-		defer cleanup()
-
-		utils.WriteAsJSON(w, postgresdebugging.Postgres(cs.Postgres), utils.GetPrettyPrintFromQueryParams(req))
-	})
-
-	httpMux.HandleFunc("/debug/redis_monitoring", func(w http.ResponseWriter, req *http.Request) {
-		if !coreconfig.SystemProbe().GetBool("service_monitoring_config.enable_redis_monitoring") {
-			writeDisabledProtocolMessage("redis", w)
-			return
-		}
-		id := utils.GetClientID(req)
-		cs, cleanup, err := nt.tracer.GetActiveConnections(id)
-		if err != nil {
-			log.Errorf("unable to retrieve connections: %s", err)
-			w.WriteHeader(500)
-			return
-		}
-		defer cleanup()
-
-		utils.WriteAsJSON(w, redisdebugging.Redis(cs.Redis), utils.GetPrettyPrintFromQueryParams(req))
-	})
-
-	httpMux.HandleFunc("/debug/http2_monitoring", func(w http.ResponseWriter, req *http.Request) {
-		if !coreconfig.SystemProbe().GetBool("service_monitoring_config.enable_http2_monitoring") {
-			writeDisabledProtocolMessage("http2", w)
-			return
-		}
-		id := utils.GetClientID(req)
-		cs, cleanup, err := nt.tracer.GetActiveConnections(id)
-		if err != nil {
-			log.Errorf("unable to retrieve connections: %s", err)
-			w.WriteHeader(500)
-			return
-		}
-		defer cleanup()
-
-		utils.WriteAsJSON(w, httpdebugging.HTTP(cs.HTTP2, cs.DNS), utils.GetPrettyPrintFromQueryParams(req))
-	})
-
 	// /debug/ebpf_maps as default will dump all registered maps/perfmaps
 	// an optional ?maps= argument could be pass with a list of map name : ?maps=map1,map2,map3
 	httpMux.HandleFunc("/debug/ebpf_maps", func(w http.ResponseWriter, req *http.Request) {
@@ -291,12 +198,7 @@ func (nt *networkTracer) Register(httpMux *module.Router) error {
 		utils.WriteAsJSON(w, cache, utils.CompactOutput)
 	})
 
-	httpMux.HandleFunc("/debug/usm_telemetry", telemetry.Handler)
-	httpMux.HandleFunc("/debug/usm/traced_programs", usm.GetTracedProgramsEndpoint(usmconsts.USMModuleName))
-	httpMux.HandleFunc("/debug/usm/blocked_processes", usm.GetBlockedPathIDEndpoint(usmconsts.USMModuleName))
-	httpMux.HandleFunc("/debug/usm/clear_blocked", usm.GetClearBlockedEndpoint(usmconsts.USMModuleName))
-	httpMux.HandleFunc("/debug/usm/attach-pid", usm.GetAttachPIDEndpoint(usmconsts.USMModuleName))
-	httpMux.HandleFunc("/debug/usm/detach-pid", usm.GetDetachPIDEndpoint(usmconsts.USMModuleName))
+	registerUSMEndpoints(nt, httpMux)
 
 	// Convenience logging if nothing has made any requests to the system-probe in some time, let's log something.
 	// This should be helpful for customers + support to debug the underlying issue.
