+++ cmd/system-probe/modules/usm_endpoints_windows.go
@@ -1,0 +1,16 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2025-present Datadog, Inc.
+
+//go:build windows && npm
+
+package modules
+
+import (
+	"github.com/DataDog/datadog-agent/pkg/system-probe/api/module"
+)
+
+func registerUSMEndpoints(nt *networkTracer, httpMux *module.Router) {
+	registerUSMCommonEndpoints(nt, httpMux)
+}
